remove forked coinstake tx

+++ electrum/address_synchronizer.py
@@ -377,11 +377,17 @@ class AddressSynchronizer(Logger):
             old_hist = self.get_address_history(addr)
             for tx_hash, height in old_hist:
                 if (tx_hash, height) not in hist:
+                    # if coinstake, just remove it
+                    tx = self.db.get_transaction(tx_hash)
+                    if tx.is_coinstake():
+                        self.remove_transaction(tx_hash)
+                        continue
                     # make tx local
                     self.unverified_tx.pop(tx_hash, None)
                     self.db.remove_verified_tx(tx_hash)
                     if self.verifier:
                         self.verifier.remove_spv_proof_for_tx(tx_hash)
+
             self.db.set_addr_history(addr, hist)
 
         for tx_hash, tx_height in hist:
